+++ screenpy_playwright/abilities/browse_the_web_synchronously.py
@@ -2,18 +2,23 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+from collections import defaultdict
+from typing import TYPE_CHECKING, Callable
 
 from playwright.sync_api import sync_playwright
 
+from ..exceptions import NoPageError
+
 if TYPE_CHECKING:
-    from typing import TypeVar
-
-    from playwright.sync_api import Browser, BrowserContext, Page, Playwright
-
-    SelfBrowseTheWebSynchronously = TypeVar(
-        "SelfBrowseTheWebSynchronously", bound="BrowseTheWebSynchronously"
+    from playwright.sync_api import (
+        Browser,
+        BrowserContext,
+        ConsoleMessage,
+        Error as PlaywrightError,
+        Page,
+        Playwright,
     )
+    from typing_extensions import Self
 
 
 class BrowseTheWebSynchronously:
@@ -33,57 +38,85 @@ class BrowseTheWebSynchronously:
     """
 
     playwright: Playwright | None = None
-    current_page: Page | None
+    _current_page: Page | None
     pages: list[Page]
+    console_logs: dict[Page, list[ConsoleMessage | PlaywrightError]]
 
     @classmethod
-    def using(
-        cls: type[SelfBrowseTheWebSynchronously],
-        playwright: Playwright,
-        browser: Browser | BrowserContext,
-    ) -> SelfBrowseTheWebSynchronously:
+    def using(cls, playwright: Playwright, browser: Browser | BrowserContext) -> Self:
         """Supply a pre-defined Playwright browser to use."""
         cls.playwright = playwright
         return cls(browser)
 
     @classmethod
-    def using_firefox(
-        cls: type[SelfBrowseTheWebSynchronously],
-    ) -> SelfBrowseTheWebSynchronously:
+    def using_firefox(cls) -> Self:
         """Use a synchronous Firefox browser."""
         if cls.playwright is None:
             cls.playwright = sync_playwright().start()
         return cls(cls.playwright.firefox.launch())
 
     @classmethod
-    def using_chromium(
-        cls: type[SelfBrowseTheWebSynchronously],
-    ) -> SelfBrowseTheWebSynchronously:
+    def using_chromium(cls) -> Self:
         """Use a synchronous Chromium (i.e. Chrome, Edge, Opera, etc.) browser."""
         if cls.playwright is None:
             cls.playwright = sync_playwright().start()
         return cls(cls.playwright.chromium.launch())
 
     @classmethod
-    def using_webkit(
-        cls: type[SelfBrowseTheWebSynchronously],
-    ) -> BrowseTheWebSynchronously:
+    def using_webkit(cls) -> Self:
         """Use a synchronous WebKit (i.e. Safari, etc.) browser."""
         if cls.playwright is None:
             cls.playwright = sync_playwright().start()
         return cls(cls.playwright.webkit.launch())
 
-    def forget(self: SelfBrowseTheWebSynchronously) -> None:
+    def __init__(self, browser: Browser | BrowserContext) -> None:
+        self.browser = browser
+        self._current_page = None
+        self.pages = []
+        self.console_logs = defaultdict(list)
+
+    @property
+    def current_page(self) -> Page:
+        """Get the current page.
+
+        Raises a :class:`~screenpy_playwright.exceptions.NoPageError` if there
+        is no current page.
+        """
+        if self._current_page is None:
+            msg = "There is no current page. Did you forget to `Open` a page?"
+            raise NoPageError(msg)
+        return self._current_page
+
+    @current_page.setter
+    def current_page(self, page: Page) -> None:
+        """Set the current page."""
+        self._current_page = page
+
+    def _add_console_log_factory(self, page: Page) -> Callable[[ConsoleMessage], None]:
+        def _add_console_log(msg: ConsoleMessage) -> None:
+            self.console_logs[page].append(msg)
+
+        return _add_console_log
+
+    def _add_page_error_factory(self, page: Page) -> Callable[[PlaywrightError], None]:
+        def _add_page_error(err: PlaywrightError) -> None:
+            self.console_logs[page].append(err)
+
+        return _add_page_error
+
+    def new_page(self) -> Page:
+        """Create a new page and make it the current page."""
+        page = self.browser.new_page()
+        console_callback = self._add_console_log_factory(page)
+        pageerror_callback = self._add_page_error_factory(page)
+        page.on("console", console_callback)
+        page.on("pageerror", pageerror_callback)
+
+        self.current_page = page
+        self.pages.append(page)
+
+        return self.current_page
+
+    def forget(self) -> None:
         """Forget everything you knew about being a playwright."""
         self.browser.close()
-        if self.playwright:
-            self.playwright.stop()
-        self.__class__.playwright = None
-
-    def __init__(
-        self: SelfBrowseTheWebSynchronously,
-        browser: Browser | BrowserContext,
-    ) -> None:
-        self.browser = browser
-        self.current_page = None
-        self.pages = []