+++ screenpy_playwright/actions/save_console_log.py
@@ -2,56 +2,59 @@
 
 from __future__ import annotations
 
-import os
+from itertools import chain
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
-from playwright.sync_api import Error as PlaywrightError
-from screenpy import AttachTheFile, DeliveryError, UnableToAct, beat
+from screenpy import AttachTheFile, DeliveryError, beat
 
 from ..abilities import BrowseTheWebSynchronously
 
 if TYPE_CHECKING:
+    from playwright.sync_api import ConsoleMessage
     from screenpy import Actor
     from typing_extensions import Self
 
 
-class SaveScreenshot:
-    """Save a screenshot of the Actor's current page.
+class SaveConsoleLog:
+    """Save the Actor's browser's console log.
 
-    Use the :meth:`~SaveScreenshot.and_attach_it` method to indicate that this
-    screenshot should be attached to all reports through the Narrator's
-    adapters. This method also accepts any keyword arguments those adapters
-    might require.
+    Use the :meth:`~SaveConsoleLog.and_attach_it` method to indicate that this
+    file should be attached to all reports through the Narrator's adapters. This
+    method also accepts any keyword arguments those adapters might require.
 
     Abilities Required:
         :class:`~screenpy_playwright.abilities.BrowseTheWebSynchronously`
 
     Examples::
 
-        the_actor.attempts_to(SaveScreenshot("screenshot.png"))
+        the_actor.attempts_to(SaveConsoleLog("consolelog.txt"))
 
-        the_actor.attempts_to(SaveScreenshot.as_(filepath))
+        the_actor.attempts_to(SaveConsoleLog.as_(filepath))
 
         # attach file to the Narrator's reports (behavior depends on adapter).
-        the_actor.attempts_to(SaveScreenshot.as_(filepath).and_attach_it())
+        the_actor.attempts_to(SaveConsoleLog.as_(filepath).and_attach_it())
 
         # using screenpy_adapter_allure plugin!
         from allure_commons.types import AttachmentTypes
         the_actor.attempts_to(
-            SaveScreenshot.as_(filepath).and_attach_it_with(
-                attachment_type=AttachmentTypes.PNG,
+            SaveConsoleLog.as_(filepath).and_attach_it_with(
+                attachment_type=AttachmentTypes.TXT,
             ),
         )
     """
 
     attach_kwargs: dict | None
-    path: str
+    path: Path
     filename: str
 
+    def describe(self) -> str:
+        """Describe the Action in present tense."""
+        return f"Save browser console log as {self.filename}"
+
     @classmethod
-    def as_(cls, path: str) -> Self:
-        """Supply the name and/or filepath for the screenshot.
+    def as_(cls, path: str | Path) -> Self:
+        """Supply the name and/or filepath for the console log.
 
         If only a name is supplied, the screenshot will appear in the current
         working directory.
@@ -61,14 +64,10 @@ class SaveScreenshot:
         """
         return cls(path=path)
 
-    def __init__(self, path: str) -> None:
-        self.path = path
-        self.filename = path.split(os.path.sep)[-1]
+    def __init__(self, path: str | Path) -> None:
+        self.path = Path(path)
+        self.filename = self.path.name
         self.attach_kwargs = None
-
-    def describe(self) -> str:
-        """Describe the Action in present tense."""
-        return f"Save screenshot as {self.filename}"
 
     def and_attach_it(self, **kwargs: Any) -> Self:  # noqa: ANN401
         """Indicate the screenshot should be attached to any reports.
@@ -87,24 +86,22 @@ class SaveScreenshot:
 
     and_attach_it_with = and_attach_it
 
-    @beat("{} saves a screenshot as {filename}")
+    @beat("{} saves their browser's console log as {filename}")
     def perform_as(self, the_actor: Actor) -> None:
-        """Direct the actor to save a screenshot."""
-        current_page = the_actor.ability_to(BrowseTheWebSynchronously).current_page
-        if current_page is None:
-            msg = "No page has been opened! Cannot save a screenshot."
-            raise UnableToAct(msg)
+        """Direct the actor to save their browser's console log."""
+        browse_the_web = the_actor.ability_to(BrowseTheWebSynchronously)
+        all_logs: chain[ConsoleMessage] = chain.from_iterable(
+            browse_the_web.console_logs.values()  # type: ignore[arg-type]
+        )
 
         try:
-            screenshot = current_page.screenshot(path=self.path)
-        except PlaywrightError as e:
+            self.path.write_text("\n".join(str(log) for log in all_logs))
+        except OSError as e:
             msg = (
                 f"{the_actor} encountered an issue while attempting to "
-                f"save a screenshot: {e.__class__.__name__}"
+                f"save their console log: {e.__class__.__name__}"
             )
             raise DeliveryError(msg) from e
 
-        Path(self.path).write_bytes(screenshot)
-
         if self.attach_kwargs is not None:
-            the_actor.attempts_to(AttachTheFile(self.path, **self.attach_kwargs))
+            the_actor.attempts_to(AttachTheFile(str(self.path), **self.attach_kwargs))